chore(lexer): Simplified lexer and organazied it
