+++ Makefile
@@ -1,13 +1,14 @@
 NAME		=	webserv
 ARGS		?=	config/example.conf
-# =============================== COMPILATION ================================ #
-CXX			=	c++
-CXXFLAGS	=	-std=c++98 -Wall -Wextra -Werror -I$(HDIR) -MMD -MP
 # =============================== DIRECTORIES ================================ #
 HDIR		=	inc
 ODIR		=	obj
 SDIR		=	src
 LOGDIR		=	log_valgrind
+# =============================== COMPILATION ================================ #
+CXX			=	c++
+CXXFLAGS	=	-std=c++98 -Wall -Wextra -Werror -MMD -MP
+CXXFLAGS	+=	-I$(HDIR) -I$(HDIR)/config
 # ================================== FILES =================================== #
 SRCS		:=	$(shell find $(SDIR) -name "*.cpp")
 OBJS		:=	$(patsubst $(SDIR)/%.cpp,$(ODIR)/%.o,$(SRCS))
